small refactor separating controllers
