+++ src/Application/UseCase/GetUserUseCase.php
@@ -3,20 +3,20 @@
 namespace App\Application\UseCase;
 
 use App\Domain\Repository\UserRepositoryInterface;
-use App\Application\Request\GetUserInformationRequest;
-use App\Application\Response\GetUserInformationResponse;
+use App\Application\Request\GetUserRequest;
+use App\Application\Response\GetUserResponse;
 
-class GetUserInformationUseCase{
+class GetUserUseCase{
     public function __construct(UserRepositoryInterface $userRepositoryInterface)
     {
         $this->userRepositoryInterface = $userRepositoryInterface;
     }
 
-    public function get(GetUserInformationRequest $request)
+    public function get(GetUserRequest $request)
     {
         $userInformation = $this->userRepositoryInterface->findById($request->getId());
          
-        return new GetUserInformationResponse($userInformation->id, $userInformation->user_name, $userInformation->email);
+        return new GetUserResponse($userInformation->id, $userInformation->user_name, $userInformation->email);
     }
 
 }
